Integrated the Mollie API monkeypatch calls as pytest fixtures
